+++ com.xpturn.klotho/Runtime/AssemblyInfo.cs
@@ -2,3 +2,4 @@ using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("xpTURN.Klotho.Runtime.Unity")]
 [assembly: InternalsVisibleTo("xpTURN.Klotho.Tests")]
+[assembly: InternalsVisibleTo("BrawlerDedicatedServer")]